+++ index.html
@@ -227,6 +227,7 @@
       </div>
     </section>
   </template>
+  <script src = './js/main.js'></script>
 
 </body>
 </html>
