+++ js/main.js
@@ -1,19 +1,9 @@
 /* eslint-disable no-unused-vars */
 import {
-  PICTURE_COUNT
-  , MIN_LIKES_COUNT
-  , MAX_LIKES_COUNT
-  , AVATAR_COUNT
-  , COMMENT_COUNT
-  , COMMENTS
-  , DESCRIPTIONS
-  , NAMES
-  , getRandomInt
-  , getRandomArrayElement
-  , createIdGenerator
-  , generateCommentId
+  generateCommentId
   , createMessage
   , createComment
   , createPicture
   , getPictures} from './data.js';
 import {lenStrCheck} from './functions.js';
+import {getRandomInt, getRandomArrElement, createID} from './util.js';
